refactor(scripts): close_track_if_done verhuist naar track_reconciler_closure.py
Fase 1 PR 2, track file-size-refactor-debt. Pure verplaatsing (345 regels,
AST-IDENTIEK via refactor_equivalence.py) met re-export op de oude plek;
geen consumer gewijzigd.

Herstelt tevens het stil gebroken ADVISORY ONLY-contract van
track_reconciler.py: het achterblijvende bestand schrijft nu aantoonbaar
alleen tracks.derived_status (1x UPDATE, geverifieerd). De nieuwe module
krijgt een eigen eerlijke docstring: schrijft tracks.phase en advanceert
tracks via tracks_lib.transition_phase, onder human-gates (actor,
approval_id, include_parked-uitzondering).

Bewijslast (plan §3.2 + zesde bewijs):
1. size-gate 1006 warning -> 667 warning (closure: 405, ok)
2. refactor_equivalence.py: IDENTIEK, exit 0
3. refactor_surface.py: OK met identiteitscontrole, exit 0
4. pytest 8 consumententests: 1 failed, 328 passed (voor == na; de ene
   failure is pre-existing OI-934)
5. git diff --stat raakt geen consumer-bestand
6. build_t0_state.py structurele diff: alleen live-eventfeed-drift
   (dispatch_register_events + 1 pr_status-entry), tracks-sectie identiek,
   NA-vs-NA-controle 0 afwijkingen

Dispatch-ID: 20260802-f1pr2-track-reconciler-closure
